Fixes bug with RepositoryService passing itself instead of client
The signature for the AtomRepository constructor has the first argument
being the cmis client, but the value that's passed by the
RepositoryService is itself, not the client. This commit fixes an issue
where any method of the repository objects that result in an HTTP
request fail with an AttributeError, due to the client not being
available.